終了状態

+++ src/recipes/recipes.service.ts
@@ -4,6 +4,8 @@ import { Repository } from 'typeorm';
 import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { UpdateRecipeDto } from './dto/update-recipe.dto';
 import Recipes from './entities/recipe.entity';
+import { plainToInstance } from 'class-transformer';
+import { RecipeDto } from './dto/recipes.dto';
 
 @Injectable()
 export class RecipesService {
@@ -12,31 +14,53 @@ export class RecipesService {
     private readonly recipesRepository: Repository<Recipes>,
   ) {}
   async create(createRecipeDto: CreateRecipeDto) {
-    const createRecipe: CreateRecipeDto = {
-      id: 111,
-      title: createRecipeDto.title,
-      making_time: createRecipeDto.making_time,
-      serves: createRecipeDto.serves,
-      ingredients: createRecipeDto.ingredients,
-      cost: createRecipeDto.cost,
-      created_at: new Date(),
-      updated_at: new Date(),
-    };
-    const recipe = await this.recipesRepository.save(createRecipe);
-    return recipe;
+    if (
+      createRecipeDto.title &&
+      createRecipeDto.making_time &&
+      createRecipeDto.serves &&
+      createRecipeDto.ingredients &&
+      createRecipeDto.cost
+    ) {
+      const createRecipe: CreateRecipeDto = {
+        title: createRecipeDto.title,
+        making_time: createRecipeDto.making_time,
+        serves: createRecipeDto.serves,
+        ingredients: createRecipeDto.ingredients,
+        cost: createRecipeDto.cost,
+        created_at: new Date(),
+        updated_at: new Date(),
+      };
+      const recipe = await this.recipesRepository.save(createRecipe);
+      return {
+        message: 'Recipe successfully created!',
+        recipes: recipe,
+      };
+    } else {
+      return {
+        message: 'Recipe creation failed!',
+        required: 'title, making_time, serves, ingredients, cost',
+      };
+    }
   }
 
   async findAll() {
-    return await this.recipesRepository.find();
+    const recipes: Recipes[] = await this.recipesRepository.find();
+    return { recipes: plainToInstance(RecipeDto, recipes) };
   }
 
   async findOne(id: number) {
-    return await this.recipesRepository.findOneBy({ id });
+    const recipe: Recipes | null = await this.recipesRepository.findOneBy({
+      id,
+    });
+    if (recipe === null) return;
+    return {
+      message: 'Recipe details by id',
+      recipes: [plainToInstance(RecipeDto, recipe)],
+    };
   }
 
   async update(id: number, updateRecipeDto: UpdateRecipeDto) {
     const updateRecipe: UpdateRecipeDto = {
-      id: id,
       title: updateRecipeDto.title,
       making_time: updateRecipeDto.making_time,
       serves: updateRecipeDto.serves,
@@ -46,15 +70,19 @@ export class RecipesService {
       updated_at: new Date(),
     };
     await this.recipesRepository.update(id, updateRecipe);
-    return this.findOne(id);
+    const recipe = this.findOne(id);
+    return {
+      message: 'Recipe successfully updated!',
+      recipes: [plainToInstance(RecipeDto, recipe)],
+    };
   }
 
   async remove(id: number) {
-    try {
-      await this.findOne(id);
+    const removeRecipe = await this.findOne(id);
+    if (removeRecipe) {
       await this.recipesRepository.delete(id);
       return { message: 'Recipe successfully removed!' };
-    } catch {
+    } else {
       return { message: 'No Recipe found' };
     }
   }
